Create authMiddleware, adminMiddleware, schema validation

+++ backend/routes/courseRoutes.js
@@ -1,12 +1,17 @@
 import express from "express";
 import Course from "../models/Course.js";
+import authMiddleware from "../middleware/authMiddleware.js";
+import adminMiddleware from "../middleware/adminMiddleware.js";
+import wrapAsync from "../utils/wrapAsync.js";
+import validateCourse from "../middleware/validateCourse.js";
+import AppError from "../utils/error.js";
 const router = express.Router();
 
-router.get("/courses", async (req, res) => {
+router.get("/courses", wrapAsynce(async (req, res) => {
     const data = await Course.find({});
     res.json(data);
-})
-router.post("/courses", async (req, res) => {
+}))
+router.post("/courses", authMiddleware, adminMiddleware, validateCourse, wrapAsync(async (req, res) => {
     const courseData = req.body;
     const newCourse = new Course({
         ...courseData
@@ -16,29 +21,38 @@ router.post("/courses", async (req, res) => {
         message: "Course saved successfully"
     });
 
-})
-router.get("/courses/:id", async(req, res) => {
+}))
+router.get("/courses/:id", wrapAsync(async (req, res) => {
     const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new AppError("Invalid ID", 400);
+    }
     const findCourse = await Course.findById(id);
     if (!findCourse) {
-        return res.json({message: "Data not found"});
+        throw new AppError("Course not found", 404);
     }
     res.json(findCourse);
-})
-router.put("/courses/:id", async (req, res) => {
+}))
+router.put("/courses/:id", authMiddleware, adminMiddleware, validateCourse, wrapAsync(async (req, res) => {
     const { id } = req.params;
     const updateValue = req.body;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new AppError("Invalid ID", 400);
+    }
     const findData = await Course.findById(id);
     if (!findData) {
-        return res.json({ message: "Data not found" });
+        throw new AppError("Data not found", 404);
     }
     await Course.findByIdAndUpdate(id, updateValue);
     res.json({
         message: "Course updated successfully"
     });
-})
-router.delete("/courses/:id", async(req, res) => {
+}))
+router.delete("/courses/:id", authMiddleware, adminMiddleware, wrapAsync(async (req, res) => {
     const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new AppError("Invalid ID", 400);
+    }
     const deletedCourse = await Course.findByIdAndDelete(id);
     if (!deletedCourse) {
         return res.json({ message: "Data not found" });
@@ -47,6 +61,6 @@ router.delete("/courses/:id", async(req, res) => {
         message: "Course deleted"
     });
 
-})
+}))
 
 export default router;
